내가 제보한 카페 확인 후 삭제

+++ controller/reportController.js
@@ -91,12 +91,13 @@ module.exports = {
         return res.status(statusCode.BAD_REQUEST).send(util.fail(statusCode.BAD_REQUEST, responseMessage.NOT_RIGHT_REPORT_USER));
       }
 
-      const result = await cafeService.deleteCafe(cafeId);
-      if (!result) {
-        return res.status(statusCode.BAD_REQUEST).send(util.fail(statusCode.BAD_REQUEST, responseMessage.DELETE_CAFE_FAIL));
+      const isNotRealYet = await cafeService.checkCafeIsNotReal()
+      if (!isNotRealYet) {
+        return res.status(statusCode.BAD_REQUEST).send(util.fail(statusCode.BAD_REQUEST, responseMessage.IS_REAL_CAFE));
       }
 
-      return res.status(statusCode.OK).send(util.success(statusCode.OK,responseMessage.DELETE_CAFE_SUCCESS, isRightReportUser));
+      const result = await cafeService.deleteCafe(cafeId);
+      return res.status(statusCode.OK).send(util.success(statusCode.OK,responseMessage.DELETE_CAFE_SUCCESS));
     } catch (error) {
       return res.status(statusCode.INTERNAL_SERVER_ERROR).send(util.fail(statusCode.INTERNAL_SERVER_ERROR,responseMessage.INTERNAL_SERVER_ERROR));
     }
